+++ src/main/resources/db/migration/V2__Create_Table_Staff.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `tb_staff` (
   `id` bigint NOT NULL AUTO_INCREMENT PRIMARY KEY,
   `avatar_url` varchar(255) DEFAULT NULL,
-  `base_salary` double DEFAULT NULL,
+  `base_salary` int DEFAULT NULL,
   `cpf` varchar(11) NOT NULL,
   `created_at` datetime(6) DEFAULT NULL,
   `email` varchar(255) NOT NULL,
